sparse_fd_list in FD
